Bug 564420 move HC condition miner to new interfaces
side work:

 - put `lic.api.tests` internal packages into nitation order
