+++ tests/org.eclipse.passage.lic.api.tests/META-INF/MANIFEST.MF
@@ -9,6 +9,6 @@ Bundle-Copyright: %Bundle-Copyright
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Require-Bundle: org.junit;bundle-version="4.12.0",
  org.eclipse.passage.lic.api;bundle-version="0.7.0"
-Export-Package: org.eclipse.passage.lic.api.conditions.tests;x-internal:=true,
- org.eclipse.passage.lic.api.tests;x-internal:=true,
- org.eclipse.passage.lic.api.version.tests;x-internal:=true
+Export-Package: org.eclipse.passage.lic.api.tests;x-internal:=true,
+ org.eclipse.passage.lic.api.tests.conditions;x-internal:=true,
+ org.eclipse.passage.lic.api.tests.version;x-internal:=true
