+++ tests/org.eclipse.passage.lic.api.tests/src/org/eclipse/passage/lic/api/tests/version/VersionContractTest.java
@@ -10,33 +10,27 @@
  * Contributors:
  *     ArSysOp - initial API and implementation
  *******************************************************************************/
-package org.eclipse.passage.lic.api.conditions.tests;
+package org.eclipse.passage.lic.api.tests.version;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import java.util.Date;
-
-import org.eclipse.passage.lic.internal.api.conditions.ValidityPeriod;
+import org.eclipse.passage.lic.internal.api.version.Version;
 import org.junit.Test;
 
 @SuppressWarnings("restriction")
-public abstract class ValidityPeriodContractTest<V extends ValidityPeriod> {
+public abstract class VersionContractTest {
 
 	@Test
-	public final void dateCanBeValid() {
-		assertTrue(atLeastMonthLongFrom(movedNow(-1)).valid(new Date()));
+	public void sameValuedVersionsAreEqual() {
+		assertTrue(fromString("1.0.0").equals(fromString("1.0.0"))); //$NON-NLS-1$//$NON-NLS-2$
 	}
 
 	@Test
-	public final void dateCanBeInvalid() {
-		assertFalse(atLeastMonthLongFrom(movedNow(1)).valid(new Date()));
+	public void differentValuedVersionsAreNotEqual() {
+		assertFalse(fromString("1.0.0").equals(fromString("1.0.1"))); //$NON-NLS-1$//$NON-NLS-2$
 	}
 
-	protected final Date movedNow(int hours) {
-		return new Date(System.currentTimeMillis() + hours * 3_600_000L);
-	}
-
-	protected abstract V atLeastMonthLongFrom(Date from);
+	protected abstract Version fromString(String source);
 
 }
